Add telemetry for new interpreters discovered
